pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Difference calculator"
authors = ["Aleksandr Pletnev <alekspletnev@gmail.com>"]
repository = "https://github.com/pletnev-aa/python-project-lvl2"
packages = [
    { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.6"

[tool.poetry.scripts]
gendiff = 'gendiff.scripts.gendiff:main'

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
pytest-cov = "^2.11.1"
PyYAML = "^5.4.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

classifiers = [
    "Programming Language :: Python :: 3.6",
]
